+++ src/Manager.php
@@ -5,9 +5,11 @@
 
 namespace Hammerstone\Sidecar;
 
+use Aws\Lambda\Exception\LambdaException;
 use Hammerstone\Sidecar\Clients\LambdaClient;
 use Hammerstone\Sidecar\Events\AfterFunctionExecuted;
 use Hammerstone\Sidecar\Events\BeforeFunctionExecuted;
+use Hammerstone\Sidecar\Exceptions\FunctionNotFoundException;
 use Hammerstone\Sidecar\Results\PendingResult;
 use Illuminate\Console\Command;
 use Illuminate\Support\Arr;
@@ -80,7 +82,7 @@ class Manager
      */
     public function getEnvironment()
     {
-        return $this->environment ?? config('app.env');
+        return $this->environment ?? config('sidecar.env') ?? config('app.env');
     }
 
     /**
@@ -105,21 +107,29 @@ class Manager
 
         $function->beforeExecution($payload);
 
-        $result = app(LambdaClient::class)->{$method}([
-            // Function name plus our alias name.
-            'FunctionName' => $function->nameWithPrefix() . ':active',
+        try {
+            $result = app(LambdaClient::class)->{$method}([
+                // Function name plus our alias name.
+                'FunctionName' => $function->nameWithPrefix() . ':active',
 
-            // `RequestResponse` is a synchronous call, vs `Event` which
-            // is a fire-and-forget, we can make it async by using the
-            // invokeAsync method.
-            'InvocationType' => 'RequestResponse',
+                // `RequestResponse` is a synchronous call, vs `Event` which
+                // is a fire-and-forget, we can make it async by using the
+                // invokeAsync method.
+                'InvocationType' => 'RequestResponse',
 
-            // Include the execution log in the response.
-            'LogType' => 'Tail',
+                // Include the execution log in the response.
+                'LogType' => 'Tail',
 
-            // Pass the payload to the function.
-            'Payload' => json_encode($payload)
-        ]);
+                // Pass the payload to the function.
+                'Payload' => json_encode($payload)
+            ]);
+        } catch (LambdaException $e) {
+            if ($e->getStatusCode() === 404) {
+                throw FunctionNotFoundException::make($function);
+            }
+
+            throw $e;
+        }
 
         // Let the calling function determine what to do with the result.
         $result = $function->toResult($result);